Split timeseries E2E suite by concern
